+++ test/gtest_unit/ts_queue.cpp
@@ -108,13 +108,13 @@ TEST(TSQueue, PushPopMulti) {
     }
 
     SList<QueueElem> slist;
-    EXPECT_EQ(slist.size(), 0);
+    EXPECT_EQ(slist.size(), 0u);
     EXPECT_TRUE(slist.empty());
     uint32_t first = 1;
     EXPECT_EQ(id_vec.front(), first);
 
     slist = q.pop_front(1);
-    EXPECT_EQ(slist.size(), 1);
+    EXPECT_EQ(slist.size(), 1u);
     EXPECT_FALSE(slist.empty());
     for (auto& elem:slist) {
         uint32_t bak_id = id_vec.front();
@@ -129,7 +129,7 @@ TEST(TSQueue, PushPopMulti) {
     EXPECT_EQ(id_vec.front(), first);
 
     slist = q.pop_front(1);
-    EXPECT_EQ(slist.size(), 1);
+    EXPECT_EQ(slist.size(), 1u);
     EXPECT_FALSE(slist.empty());
     for (auto& elem:slist) {
         uint32_t bak_id = id_vec.front();
@@ -142,7 +142,7 @@ TEST(TSQueue, PushPopMulti) {
     EXPECT_EQ(id_vec.front(), first);
 
     slist = q.pop_front(2);
-    EXPECT_EQ(slist.size(), 2);
+    EXPECT_EQ(slist.size(), 2u);
     EXPECT_FALSE(slist.empty());
     for (auto& elem:slist) {
         uint32_t bak_id = id_vec.front();
@@ -155,7 +155,7 @@ TEST(TSQueue, PushPopMulti) {
     EXPECT_EQ(id_vec.front(), first);
 
     slist = q.pop_front(3);
-    EXPECT_EQ(slist.size(), 3);
+    EXPECT_EQ(slist.size(), 3u);
     EXPECT_FALSE(slist.empty());
     for (auto& elem:slist) {
         uint32_t bak_id = id_vec.front();
@@ -168,7 +168,7 @@ TEST(TSQueue, PushPopMulti) {
     EXPECT_EQ(id_vec.front(), first);
 
     slist = q.pop_front(94);
-    EXPECT_EQ(slist.size(), 94);
+    EXPECT_EQ(slist.size(), 94u);
     EXPECT_FALSE(slist.empty());
     for (auto& elem:slist) {
         uint32_t bak_id = id_vec.front();
@@ -182,7 +182,7 @@ TEST(TSQueue, PushPopMulti) {
     EXPECT_EQ(id_vec.front(), first);
 
     slist = q.pop_front(100);
-    EXPECT_EQ(slist.size(), 100);
+    EXPECT_EQ(slist.size(), 100u);
     EXPECT_FALSE(slist.empty());
     EXPECT_TRUE(q.empty());
     for (auto& elem:slist) {
@@ -195,7 +195,7 @@ TEST(TSQueue, PushPopMulti) {
     EXPECT_EQ(id_vec.front(), first);
 
     slist = q.pop_front(101);
-    EXPECT_EQ(slist.size(), 100);
+    EXPECT_EQ(slist.size(), 100u);
     EXPECT_FALSE(slist.empty());
     EXPECT_TRUE(q.empty());
     for (auto& elem:slist) {
@@ -222,72 +222,72 @@ TEST(TSQueue, LFLock) {
 TEST(TSSkipQueueT, pushpop)
 {
     QueueElem::s_id = 0;
-    int n = 10000;
+    size_t n = 10000;
     TSSkipQueue<QueueElem, true, 64> q;
 
     // push
     q.push(new QueueElem);
     EXPECT_TRUE(q.skip_layer_.indexs_.empty());
-    GTEST_ASSERT_EQ(q.size(), 1);
-    EXPECT_EQ(q.skip_layer_.head_offset_, 0);
-    EXPECT_EQ(q.skip_layer_.tail_offset_, 1);
+    GTEST_ASSERT_EQ(q.size(), 1u);
+    EXPECT_EQ(q.skip_layer_.head_offset_, 0u);
+    EXPECT_EQ(q.skip_layer_.tail_offset_, 1u);
 
-    for (int i = 1; i <= 62; ++i)
+    for (size_t i = 1; i <= 62; ++i)
     {
         q.push(new QueueElem);
         EXPECT_TRUE(q.skip_layer_.indexs_.empty());
         GTEST_ASSERT_EQ(q.size(), i + 1);
-        EXPECT_EQ(q.skip_layer_.head_offset_, 0);
+        EXPECT_EQ(q.skip_layer_.head_offset_, 0u);
         EXPECT_EQ(q.skip_layer_.tail_offset_, i + 1);
     }
 
     q.push(new QueueElem);
-    EXPECT_EQ(q.skip_layer_.indexs_.size(), 1);
-    GTEST_ASSERT_EQ(q.size(), 64);
-    EXPECT_EQ(q.skip_layer_.head_offset_, 0);
-    EXPECT_EQ(q.skip_layer_.tail_offset_, 0);
+    EXPECT_EQ(q.skip_layer_.indexs_.size(), 1u);
+    GTEST_ASSERT_EQ(q.size(), 64u);
+    EXPECT_EQ(q.skip_layer_.head_offset_, 0u);
+    EXPECT_EQ(q.skip_layer_.tail_offset_, 0u);
 
     q.push(new QueueElem);
-    EXPECT_EQ(q.skip_layer_.indexs_.size(), 1);
-    GTEST_ASSERT_EQ(q.size(), 65);
-    EXPECT_EQ(q.skip_layer_.head_offset_, 0);
-    EXPECT_EQ(q.skip_layer_.tail_offset_, 1);
+    EXPECT_EQ(q.skip_layer_.indexs_.size(), 1u);
+    GTEST_ASSERT_EQ(q.size(), 65u);
+    EXPECT_EQ(q.skip_layer_.head_offset_, 0u);
+    EXPECT_EQ(q.skip_layer_.tail_offset_, 1u);
 
-    for (int i = 66; i <= n; ++i)
+    for (size_t i = 66; i <= n; ++i)
     {
         q.push(new QueueElem);
-        EXPECT_EQ(q.skip_layer_.indexs_.size(), i / 64);
+        EXPECT_EQ(q.skip_layer_.indexs_.size(), i / 64u);
         GTEST_ASSERT_EQ(q.size(), i);
-        EXPECT_EQ(q.skip_layer_.head_offset_, 0);
-        EXPECT_EQ(q.skip_layer_.tail_offset_, i % 64);
+        EXPECT_EQ(q.skip_layer_.head_offset_, 0u);
+        EXPECT_EQ(q.skip_layer_.tail_offset_, i % 64u);
     }
 
     // pop
     SList<QueueElem> s = q.pop_front(1);
-    EXPECT_EQ(s.begin()->id_, 1);
-    EXPECT_EQ(q.skip_layer_.indexs_.size(), n / 64);
+    EXPECT_EQ(s.begin()->id_, 1u);
+    EXPECT_EQ(q.skip_layer_.indexs_.size(), n / 64u);
     GTEST_ASSERT_EQ(q.size(), n - 1);
-    EXPECT_EQ(q.skip_layer_.head_offset_, 1);
-    EXPECT_EQ(q.skip_layer_.tail_offset_, n % 64);
+    EXPECT_EQ(q.skip_layer_.head_offset_, 1u);
+    EXPECT_EQ(q.skip_layer_.tail_offset_, n % 64u);
 
     s = q.pop_front(62);
-    int c = 2;
+    size_t c = 2;
     for (auto &elem : s)
     {
         EXPECT_EQ(elem.id_, c);
         c++;
     }
-    EXPECT_EQ(q.skip_layer_.indexs_.size(), n / 64);
+    EXPECT_EQ(q.skip_layer_.indexs_.size(), n / 64u);
     GTEST_ASSERT_EQ(q.size(), n - 63);
-    EXPECT_EQ(q.skip_layer_.head_offset_, 63);
-    EXPECT_EQ(q.skip_layer_.tail_offset_, n % 64);
+    EXPECT_EQ(q.skip_layer_.head_offset_, 63u);
+    EXPECT_EQ(q.skip_layer_.tail_offset_, n % 64u);
 
     s = q.pop_front(1);
     EXPECT_EQ(s.begin()->id_, c);
     c++;
     GTEST_ASSERT_EQ(q.size(), n - 64);
     EXPECT_EQ(q.skip_layer_.indexs_.size(), n / 64 - 1);
-    EXPECT_EQ(q.skip_layer_.head_offset_, 0);
+    EXPECT_EQ(q.skip_layer_.head_offset_, 0u);
     EXPECT_EQ(q.skip_layer_.tail_offset_, n % 64);
 
     s = q.pop_front(n);
@@ -298,13 +298,13 @@ TEST(TSSkipQueueT, pushpop)
         c++;
     }
     EXPECT_TRUE(q.skip_layer_.indexs_.empty());
-    GTEST_ASSERT_EQ(q.size(), 0);
-    EXPECT_EQ(q.skip_layer_.head_offset_, 0);
-    EXPECT_EQ(q.skip_layer_.tail_offset_, 0);
+    GTEST_ASSERT_EQ(q.size(), 0u);
+    EXPECT_EQ(q.skip_layer_.head_offset_, 0u);
+    EXPECT_EQ(q.skip_layer_.tail_offset_, 0u);
 
-    for (int i = 0; i < 100; ++i)
+    for (size_t i = 0; i < 100; ++i)
     {
-        for (int j = 0; j < n; ++j)
+        for (size_t j = 0; j < n; ++j)
         {
             QueueElem *p = new QueueElem(j);
             GTEST_ASSERT_EQ(p->id_, j);
@@ -312,8 +312,8 @@ TEST(TSSkipQueueT, pushpop)
         }
         GTEST_ASSERT_EQ(q.size(), n);
         s = q.pop_front(-1);
-        GTEST_ASSERT_EQ(q.size(), 0);
-        int j = 0;
+        GTEST_ASSERT_EQ(q.size(), 0u);
+        size_t j = 0;
         for (auto &elem : s)
         {
             EXPECT_EQ(elem.id_, j);
@@ -452,34 +452,34 @@ TEST(Ref, TsQueueRef)
 {
     {
         TestRef * ptr = new TestRef;
-        EXPECT_EQ(TestRef::s_count, 1);
-        EXPECT_EQ(ptr->reference_, 1);
+        EXPECT_EQ(TestRef::s_count, 1u);
+        EXPECT_EQ(ptr->reference_, 1u);
         TSQueue<TestRef> q;
         q.push(ptr);
-        EXPECT_EQ(ptr->reference_, 2);
+        EXPECT_EQ(ptr->reference_, 2u);
         TestRef * out = q.pop();
-        EXPECT_EQ(ptr->reference_, 1);
+        EXPECT_EQ(ptr->reference_, 1u);
         out->DecrementRef();
-        EXPECT_EQ(TestRef::s_count, 0);
+        EXPECT_EQ(TestRef::s_count, 0u);
     }
 
     {
         TestRef * ptr = new TestRef;
-        EXPECT_EQ(TestRef::s_count, 1);
-        EXPECT_EQ(ptr->reference_, 1);
+        EXPECT_EQ(TestRef::s_count, 1u);
+        EXPECT_EQ(ptr->reference_, 1u);
         TSQueue<TestRef> q;
         q.push(ptr);
-        EXPECT_EQ(ptr->reference_, 2);
+        EXPECT_EQ(ptr->reference_, 2u);
         q.erase(ptr);
-        EXPECT_EQ(ptr->reference_, 1);
+        EXPECT_EQ(ptr->reference_, 1u);
         ptr->DecrementRef();
-        EXPECT_EQ(TestRef::s_count, 0);
+        EXPECT_EQ(TestRef::s_count, 0u);
     }
 
     {
-        const int n = 128;
+        const size_t n = 128;
         TSQueue<TestRef> q;
-        for (int i = 0; i < n; ++i)
+        for (size_t i = 0; i < n; ++i)
             q.push(new TestRef);
         EXPECT_EQ(TestRef::s_count, n);
         SList<TestRef> slist = q.pop_all();
@@ -488,13 +488,13 @@ TEST(Ref, TsQueueRef)
             elem.DecrementRef();
         EXPECT_EQ(TestRef::s_count, n);
         slist.clear();
-        EXPECT_EQ(TestRef::s_count, 0);
+        EXPECT_EQ(TestRef::s_count, 0u);
     }
 
     {
-        const int n = 128;
+        const size_t n = 128;
         TSQueue<TestRef> q;
-        for (int i = 0; i < n; ++i)
+        for (size_t i = 0; i < n; ++i)
             q.push(new TestRef);
         EXPECT_EQ(TestRef::s_count, n);
         SList<TestRef> slist = q.pop_front(28);
@@ -503,7 +503,7 @@ TEST(Ref, TsQueueRef)
             elem.DecrementRef();
         EXPECT_EQ(TestRef::s_count, n);
         slist.clear();
-        EXPECT_EQ(TestRef::s_count, 100);
+        EXPECT_EQ(TestRef::s_count, 100u);
         q.push(q.pop_all());
         q.push(q.pop_all());
         q.push(q.pop_all());
@@ -511,16 +511,16 @@ TEST(Ref, TsQueueRef)
         slist = q.pop_all();
         for (auto &elem : slist)
             elem.DecrementRef();
-        EXPECT_EQ(TestRef::s_count, 100);
+        EXPECT_EQ(TestRef::s_count, 100u);
     }
 }
 
 TEST(Ref, TsSkipQueueRef)
 {
     {
-        const int n = 128;
+        const size_t n = 128;
         TSSkipQueue<TestRef> q;
-        for (int i = 0; i < n; ++i)
+        for (size_t i = 0; i < n; ++i)
             q.push(new TestRef);
         EXPECT_EQ(TestRef::s_count, n);
         SList<TestRef> slist = q.pop_front(128);
@@ -529,13 +529,13 @@ TEST(Ref, TsSkipQueueRef)
             elem.DecrementRef();
         EXPECT_EQ(TestRef::s_count, n);
         slist.clear();
-        EXPECT_EQ(TestRef::s_count, 0);
+        EXPECT_EQ(TestRef::s_count, 0u);
     }
 
     {
-        const int n = 128;
+        const size_t n = 128;
         TSSkipQueue<TestRef> q;
-        for (int i = 0; i < n; ++i)
+        for (size_t i = 0; i < n; ++i)
             q.push(new TestRef);
         EXPECT_EQ(TestRef::s_count, n);
         SList<TestRef> slist = q.pop_front(28);
@@ -544,11 +544,11 @@ TEST(Ref, TsSkipQueueRef)
             elem.DecrementRef();
         EXPECT_EQ(TestRef::s_count, n);
         slist.clear();
-        EXPECT_EQ(TestRef::s_count, 100);
+        EXPECT_EQ(TestRef::s_count, 100u);
         slist = q.pop_front(100);
         for (auto &elem : slist)
             elem.DecrementRef();
-        EXPECT_EQ(TestRef::s_count, 100);
+        EXPECT_EQ(TestRef::s_count, 100u);
     }
 }
